+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # UNRELEASED
 
+-   Add `OlmMachine.registerDevicesUpdatedCallback` to notify when devices have
+    been updated.
+    ([#88](https://github.com/matrix-org/matrix-rust-sdk-crypto-wasm/pull/88))
+
 # matrix-sdk-crypto-wasm v4.0.1
 
 -   `PickledInboundGroupSession.sender_signing_key` is now optional.